Move user answers button into other operations dropdown
